Add MuchResult::Config/CallMethod/WrapMethod/WrapAndCallMethod
This adds `MuchRails::Config`, `MuchRails::CallMethod`,
`MuchRailsWrapMethod` and `MuchRails::WrapAndCallMethod`.
Descriptions for each mix-in are included in the mix-in itself.
This is all prep for setting up `MuchRails`.

I updated the gemspec to include `active_support` in order to use
its `classify` method in `MuchRails::Config`. However, I brought
in its core extensions library in anticipation of utilizing more
of it's methods.

This also brings in the latest versions of `MuchPlugin` and
`MuchResult`.

Lastly, this also drops the Ruby version from 2.5.5 to 2.5.3. There
were issues getting 2.5.5 installed on my current machine. As time
allows, I'll work on resolving this issue.