bring in Dassets; add Railtie that configures/initializes Dassets
This aliases Dassets as MuchRails::Assets as is the convention
for dependencies that MuchRails brings in. This also adds a Railtie
and configuration method for configuring Dassets in a default
manner. This is an opinionated default. In the future, we'll most
likely add something like profiles or provide a way to manually
configure source behavior.