move error handling into MuchRails::Action::UnprocessableEntityResult
Before this was just relying on the controller to implement a
`action_errors_json` method. This was only passing tests b/c of
stubbing and test doubles.

This switches to do the error handling in the Result directly
since it is the only thing that needs it right now. This also
updates the tests to formally test the behavior with a more
accurate controller test double.