add MuchRails::RailsRoutes
This is a convenience Singleton to access Rails' url helpers.
This will be used to generate paths/URLs from the Action router
which is coming.

Note: I had to bring in Rails as a development dependency so I
could test this. I did the bare minimum to configure it for
testing in the test helper.