fix routing to expect URL _names_ or paths
Previously, I had tested and implemented such that routing expected
either a path or a URL _object_. This is wrong. Routing should
expecct either a path or a URL _name_.

This reworks the implementation and updates the tests. This
switches to a `fetch` style API so we can easily fallback to
dynamic behavior when there is no URL for a given name. In routing,
we want to build an unnamed URL. In path generation, we want to
raise an exception to fail fast.