move gem configuration to nested configs on MuchRails module
This uses the new behaviors in MuchRails::Config to setup nested
configuration objects on the MuchRails module and moves disparate
config APIs into this. This sets up `MuchRails.config`,
`MuchRails.config.action`, `MuchRails.config.layout`.

In addition this moves the layout separators from default attribute
values to _configured_ defaults that can be overridden in an
initializer.