make Action `render` properly render view templates
This uses `.to_s.tableize.singularize` to turn the Action's class
into a path string and defaults the rendered template to that. This
means that, by default, Rails will look for templates in the same
folder structure/name as the Action.

This also updates the arguments that `render` takes so you can use
the Rails `render` API to override this default and specify an
ad-hoc template string.