bring in Rubocop
This brings in the Rubocop gem. This also includes a
`.rubocop.yml` which configures cop behavior, or disables a cop
based on what we believe to be acceptable violations. This yaml
inherits from a `.rubocop_stripe.yml`, which can be found here:
https://github.com/stripe/stripe-ruby/blob/master/.rubocop.yml
This was added as their conventions are a good starting point,
however, anything defined here can (and is in some cases) be
overwritten in our `.rubocop.yml`.

Note: This also updates Ruby to the latest 2.5 release, bumping
up from 2.5.3 to 2.5.8.