switch to redding/styleguide Rubocop config
This one builds off of Shopify's style guide and refines it
further for our conventions. This also switches to loading the
config from a remote URL.

I enabled a bunch of cops I feel should be enabled and corrected
the code as well. Look over these corrections and make sure you
are OK with them.

In a future commit, I'm going to extract a Rubygem to bring in the
dependency. I don't like this remote URL stuff - it is hard to
manage and we have bundler to manage dependencies for us.