Make `DestroyService` failure result similar to `SaveService` failure result
This updates the failure result that the `DestroyService` builds
to be similar to the failure result that the `SaveService` builds.
This is just to make them similar so we have the same options
when handling either failure result.

This also updates the `MuchRails::Records::DestructionInvalid`
exception to support the `DestroyService`. This adds error
full messages to the exception and renames the
`destruction_errors` to `errors`.