Allow configuring Save and Destroy service exception handling
This updates the `SaveService` and `DestroyService` mixins to
allow configuring which exceptions they rescue and how they
handle them. The goal of this is to allow configuring custom
exceptions and have them return failure (or success if you want)
results in an expected format. The original goal was to just
configure exception classes and use the same logic to build a
similar failure result but I realized that since we don't control
exceptions like `ActiveRecord::RecordInvalid` this doesn't scale
well. Thus, this allows configuring an exception class with a
block that builds the result the `SaveService` or `DestroyService`
should return.

By default, `SaveService` is configured to work with
`ActiveRecord::RecordInvalid` and `DestroyService` is configured
to work with `MuchRails::Records::DestructionInvalid`. This
matches how they previously worked before this commit.