update Action `render` method to set itself on any supported view models
This introduces an ah-hoc method for letting view models tell the
Action instances that they are interested in collaborating with
the Action. If the view model instances respond to
`#much_rails_action=` then the Action messages the view model
passing itself. The view model can then use the Action instance
however it sees fit.

Note, view models could opt to, instead, just take the action
instance or one it its properties as an argument to its
constructor. This works if the view model directly knows it wants
the Action instance. However, in some cases the View Model may
use e.g. a layout mixin that want's to message the Action. In this
case the view model doesn't know/care that something it mixes in
wants/needs the Action instancd. Why should the view model junk
up its constructor in non-obvious ways. This approach allows the
mixin that needs the Action to also define the `much_rails_action=`
method to give itself access to the Action and the main view model
is happily none-the-wiser.