add API for interacting with the Rails controller session on Actions
This removes the previous `current_session` keyword argument that
was intended to allow injecting user session objects. It is
replaces with a private `controller_session` method that is the
Rails controller instance's session. This allows for an easier
and more extendable API for interacting with session data.

This also adds a public `controller_session=` writer to ease
injecting fake session objects when unit testing actions.