update to the latest dependencies
This ensures we stay up to date with the latest stuff. This also
updates to the latest Ruby 2.7 version as 2.5.8 is no longer the
latest 2.5 version and 2.5 is pretty dated.

Note: when installing Rails, I ran into the weirdness installing
the latest mimemagic gem version. `brew install shared-mime-info`
allowed me to install Rails.