switch to composing MuchConfig
I've extracted the MuchRails::Config logic into MuchConfig so it
can be used outside of a Rails context. This switches to using the
logic in an aliased form similar to other parts of MuchRails
(e.g. MuchRails::Result). This keeps the logic DRY and flexible in
how it can be used.