add MuchRails::Layout.external_javascript handling
This matches the `.javascript` handling but creates a separate
(but optional) "bucket" for defining javascripts that aren't
internal to the application and don't need MuchRails::Assets
handling in their script tag markup.

E.g. you could use this by adding something like this to the
`application.html.erb` layout template:

```html
...
  <head>
    ...

    <% @view.external_javascripts.each do |external_javascript| %>
      <script src="<%= external_javascript %>"></script>
    <% end %>
    <% @view.javascripts.each do |javascript| %>
      <script src="<%= MuchRails::Assets[javascript].url %>" type="module"></script>
    <% end %>

    ...
  </head>
...
```

This is all optional and isn't required to use MuchRails. But this
adds the API and makes it available as needed.