host CSS/JS assets out of the app/views folder
This allows you to store CSS/JS that supports your view templates
right alongside the templates themselves (not in a completely
separate folder). This not only juxtaposes the logic, but ensures
the assets are hosted in a logic path that matches the resources
they support.

Note: I added restrictive filtering to only support CSS/JS for now.
My thinking is that only CSS/JS needs to directly map and be
juxtaposed to templates. Images and "app" CSS/JS are global to the
entire app and can be hosted out of the `app/assets` folder as
before. This restrictive filtering also ensures MuchRails::Assets
won't unintentionally host e.g. `resource/show.js.erb` Rails
templates.